requirements.txt
fastapi
uvicorn
python-dotenv
kiteconnect
requests
pyotp
pandas
numpy